refactored options
